An enjoyable afternoon learning neon intrinsics
Arm intrinsics much nicer than sse was, but getting stuff in and out
of the register set remains a real pita. I resorted to assembler
the last time I tried this stuff, but it was useful to read
up on how everything worked.

Totally non-working code, just giving up for the night.