I am just never going to make progress without checking in broken code
This exercise in part was born because I was having such trouble
elsewhere. It turns out that despite my dreams of rolling a lib
to handle ipv6 better, that there appears to be a kernel netlink
bug *in the first place* messing multiple daemons up

see rabeld's repo for more details. I finally got it reproducable
in under 10 minutes.

Also some profiling of babeld (after some major tuning),
ended up looking as though trying to optimize for neon, etc, was a win -
modern compilers WILL do a memcmp in xmms if you let them...

Which is good because I got stuck on merely loading and storing values
from that arch.