Yes it is still possible to reserve registers globally
Here we are in a world where we *are* regularly building the
entire toolchain, every day, and yet we don't do crazy stuff
like this.

The compiler claimed these were callee saved. But so far as
I know we do no floating point anywhere in babel. Not once.

So it might (hah!) be safe to do all these comparisons

OR, worst case - we compile our entire libc specially... :)

Which, um, used to be hard, but we do it all the time. Every day.

So far as I know the neon regs are correctly spilled in the kernel,
so all we have to do is wrap some syscalls in libc, and we're done.

(famous last words)