Tried to optimize the kill_the_martians routine
This is correct, I think, but longer than the original by 18
instructions.

I can get rid of the explicit true/false things, but honestly I
am afraid that the compiler turns out smarter than I am.

In particular there's this magic thing the compiler found that
I totally don't get, which jimps it to a return....

   2:	83 fe 07             	cmp    $0x7,%esi

Anyway, off to write the cycle counter.