introduce stagger.h header
I've written this so many different ways as a primitive so many times
it's hard to figure out the best way to do it again.

I should really code up the c version as a command line tool again.
it was really handy in the late 90s where I'd do a

pdsh -g all_servers stagger restart_server

The concept here is different, in that I have one or more threads
running (in a conventional arch), or a really fast way to multicast
to a ton of processors (parallella or fpga).