Refactoring interfaces a bit.
At least by factoring out the logical divisions within the struct I can
begin to see what to do with it.

I was also struck as I was doing this how much it felt like doing index
optimization for a database like interbase - reducing things not just to
codd form but for optimal storage on disk.

That said, thinking about the kinds of inputs interfaces have, and
outputs, are probably going to result in a much more massive re-org.