Notes for the day
I am most creative at night, but sloppy.

Pedantic in the morning.

Still... my pedantic goal this morning was to hook up some libs
to actually start looking useful. And all I did was reorg all
the sources.

Things are starting to evolve to a clear separation between
erm and it's clients.