More shuffling of deck chairs
somewhere in here I need to settle on a consistent style.

My own style (from too many 24 row terminals in the 80s) was incredibly
compressed - 2 character indent, lots of stuff on the same line. I'm
willing to reform, and enforce something on myself, but to what? Linux
kernel style seems reasonable, but not compressed enough for my taste.

There are three separate styles in this code now, and emacs is fighting
me on all of them.