Get my egocentric routing machine out of the way
I was thinking about this wrong.

Anything that uses erm has to be specially modified to use it.

erm itself has no (well, very few) intended dependencies on the c
library or kernel, and while you can look at any given thread running on
the machine as a "client of that machine", it really is an outward
facing server process, and should be thought of that way.

erm is a library for a very specialized computing environment. That's
all.

So, I renamed the stuff I was calling clients, to servers, keeping
the more speculative rants in the original clients dir - which will
probably move to "rants" or something like that until code emerges.