more bad assembly
I am spending more time writing short instruction sequences than
anything else. Anyway, the idea for a random thing is that it doesn't
need to be that random, it just needs to be fast.

So getting the bottom cycles only is just fine.

That said, there's something else wrong with the code thus far.