Several use cases for cycle timers
- measurement: must use a lfence or mfence instruction
- 1 semi-random number: just leverage the 32 bit cycle timer mod a small value
- 1 bigger random number: grab something random from the entropy pool
- a bunch of random numbers - leverage lfence/mfence, cycle timer, and
  entropy pool.

Or do something else entirely.

The code is almost exactly what I want, except that I want to be pulling
the random location start point off of a register.... (or maybe not)

  400ce0:       0f 31                   rdtsc
  400ce2:       c1 e0 02                shl    $0x2,%eax
  400ce5:       25 fc 00 00 00          and    $0xfc,%eax
  400cea:       44 8b 88 c0 20 60 00    mov    0x6020c0(%rax),%r9d

But the compiler complains about it.

rng.c:73:5: warning: optimization may eliminate reads and/or writes to register variables [-Wvolatile-register-var]

So much work for 4 instructions.