A bunch of notes about sse and clocks and symbols
It turns out a modern C preprocessor can cope with utf-8 and (with a
switch), handle it in code as well. So I've long ranted about not being
able to represent math directly in a given computer language, and went
looking for symbols I'd like to use more. Examples:

⊎ multiset union
⊕ booleans ⊖

Then I started wanting to rant about having to write code that
had to solve basic algebra (when you should be able to just
express it once and have the compiler solve for whichever
term you needed) - and called it a weekend.