some basic notes on the linker and some regs on the x86_64
I'm really hung up on how to deal with the linker. I know I managed
to "win" against it last time, but it did require quite a bit of scripting
and code generation to get right.

On the x86_64 front:

I am actually sort of, kind of, missing, segment registers.

I never thought I'd say that.